requirements.txt
discord.py==1.7.3
Flask>=2.2
python-dotenv==0.19.0
